requirements.txt
PyGObject
PyYAML